[TASK] Move tx_multicolumn_tt_content_drawItem to PageLayoutViewHook
