+++ composer.json
@@ -31,7 +31,6 @@
       "IchHabRecht\\Multicolumn\\": "Classes/"
     },
     "classmap": [
-      "hooks/",
       "lib/",
       "pi1/",
       "pi_sitemap/"
